+++ RSHitChance2/src/equipment/Armor.java
@@ -15,7 +15,7 @@ public class Armor extends Equipment implements ArmorInterface{
     
     private double damageReduction;//0 for non-tank armor, example values: level 90 shields: 0.09, raids armor: 0.018
     
-    public Armor(String name, String imageName, Slot slot, DefensiveCombatStyle combatStyle, int level, int damage, int armor, Buff buff,double damageReduction) {
+    public Armor(String name, String imageName, Slot slot, DefensiveCombatStyle combatStyle, int level, double damage, double armor, Buff buff,double damageReduction) {
         super(name, imageName, slot, combatStyle, level, damage, armor, buff);
         this.damageReduction = damageReduction;
     }
@@ -41,8 +41,8 @@ public class Armor extends Equipment implements ArmorInterface{
             Slot slot = Slot.valueOf(strArray[2]);
             DefensiveCombatStyle combatStyle = (DefensiveCombatStyle) CombatStyleFlyweight.getCombatStyle(strArray[3]);
             int level = Integer.parseInt(strArray[4]);
-            int damage = (int) Double.parseDouble(strArray[5]);
-            int armor = (int) Double.parseDouble(strArray[6]);
+            double damage = Double.parseDouble(strArray[5]);
+            double armor = Double.parseDouble(strArray[6]);
             double damageReduction = Double.parseDouble(strArray[7]);
             String buffString = strArray[8];
             Buff buff = buffString.equals("None")?NullBuff.getInstance():BuffFlyweight.getBuff(BuffName.valueOf(buffString));
